Renaming of the segmentation and sensitive detectors directories
