+++ CMakeLists.txt
@@ -49,7 +49,7 @@ message ( STATUS "ROOT_VERSION: ${ROOT_VERSION}" )
 find_package( Geant4 REQUIRED ) 
 find_package( LCIO REQUIRED)
 
-add_subdirectory(detectorSegmentations)
+add_subdirectory(detectorSegmentations SYSTEM )
 add_subdirectory(./detector/DetCommon)
 add_subdirectory(sensitiveDetectors)
 
@@ -83,7 +83,7 @@ target_include_directories(${PackageName}G4 PRIVATE ${PROJECT_SOURCE_DIR}/detect
 target_include_directories(${PackageName}   SYSTEM PUBLIC ${LCIO_INCLUDE_DIRS})
 target_include_directories(${PackageName}G4 SYSTEM PUBLIC ${LCIO_INCLUDE_DIRS})
 
-target_link_libraries(${PackageName}   DD4hep::DDCore DD4hep::DDRec DD4hep::DDParsers ${ROOT_LIBRARIES} ${LCIO_LIBRARIES} DetSegmentation)
+target_link_libraries(${PackageName}   DD4hep::DDCore DD4hep::DDRec DD4hep::DDParsers ${ROOT_LIBRARIES} ${LCIO_LIBRARIES} detectorSegmentations)
 target_link_libraries(${PackageName}G4 DD4hep::DDCore DD4hep::DDRec DD4hep::DDParsers DD4hep::DDG4 ${ROOT_LIBRARIES} ${Geant4_LIBRARIES} ${LCIO_LIBRARIES})
 
 #Create this_package.sh file, and install
