+++ detectorSegmentations/include/detectorSegmentations/GridEtaHandle.h
@@ -1,8 +1,8 @@
-#ifndef DD4HEP_DDCORE_GRIDPHIETA_H
-#define DD4HEP_DDCORE_GRIDPHIETA_H 1
+#ifndef DD4HEP_DDCORE_GRIDETA_H
+#define DD4HEP_DDCORE_GRIDETA_H 1
 
 // FCCSW
-#include "DetSegmentation/FCCSWGridPhiEta.h"
+#include "detectorSegmentations/GridEta.h"
 
 // DD4hep
 #include "DD4hep/Segmentations.h"
@@ -20,7 +20,7 @@ template <typename T>
 class SegmentationWrapper;
 
 /// We need some abbreviation to make the code more readable.
-typedef Handle<SegmentationWrapper<DDSegmentation::FCCSWGridPhiEta>> FCCSWGridPhiEtaHandle;
+typedef Handle<SegmentationWrapper<DDSegmentation::GridEta>> GridEtaHandle;
 
 /// Implementation class for the grid phi-eta segmentation.
 /**
@@ -42,27 +42,27 @@ typedef Handle<SegmentationWrapper<DDSegmentation::FCCSWGridPhiEta>> FCCSWGridPh
  *  \author  A. Zaborowska
  *  \version 1.0
  */
-class FCCSWGridPhiEta : public FCCSWGridPhiEtaHandle {
+class GridEta : public GridEtaHandle {
 public:
   /// Defintiion of the basic handled object
-  typedef FCCSWGridPhiEtaHandle::Object Object;
+  typedef GridEtaHandle::Object Object;
 
 public:
   /// Default constructor
-  FCCSWGridPhiEta() = default;
+  GridEta() = default;
   /// Copy constructor
-  FCCSWGridPhiEta(const FCCSWGridPhiEta& e) = default;
+  GridEta(const GridEta& e) = default;
   /// Copy Constructor from segmentation base object
-  FCCSWGridPhiEta(const Segmentation& e) : Handle<Object>(e) {}
+  GridEta(const Segmentation& e) : Handle<Object>(e) {}
   /// Copy constructor from handle
-  FCCSWGridPhiEta(const Handle<Object>& e) : Handle<Object>(e) {}
+  GridEta(const Handle<Object>& e) : Handle<Object>(e) {}
   /// Copy constructor from other polymorph/equivalent handle
   template <typename Q>
-  FCCSWGridPhiEta(const Handle<Q>& e) : Handle<Object>(e) {}
+  GridEta(const Handle<Q>& e) : Handle<Object>(e) {}
   /// Assignment operator
-  FCCSWGridPhiEta& operator=(const FCCSWGridPhiEta& seg) = default;
+  GridEta& operator=(const GridEta& seg) = default;
   /// Equality operator
-  bool operator==(const FCCSWGridPhiEta& seg) const { return m_element == seg.m_element; }
+  bool operator==(const GridEta& seg) const { return m_element == seg.m_element; }
   /// determine the position based on the cell ID
   inline Position position(const CellID& id) const { return Position(access()->implementation->position(id)); }
 
@@ -74,46 +74,30 @@ public:
   /// access the grid size in eta
   inline double gridSizeEta() const { return access()->implementation->gridSizeEta(); }
 
-  /// access the grid size in Phi
-  inline int phiBins() const { return access()->implementation->phiBins(); }
-
   /// access the coordinate offset in eta
   inline double offsetEta() const { return access()->implementation->offsetEta(); }
-
-  /// access the coordinate offset in Phi
-  inline double offsetPhi() const { return access()->implementation->offsetPhi(); }
 
   /// set the coordinate offset in eta
   inline void setOffsetEta(double offset) const { access()->implementation->setOffsetEta(offset); }
 
-  /// set the coordinate offset in Phi
-  inline void setOffsetPhi(double offset) const { access()->implementation->setOffsetPhi(offset); }
-
   /// set the grid size in eta
   inline void setGridSizeEta(double cellSize) const { access()->implementation->setGridSizeEta(cellSize); }
-
-  /// set the grid size in Phi
-  inline void setPhiBins(int cellSize) const { access()->implementation->setPhiBins(cellSize); }
 
   /// access the field name used for eta
   inline const std::string& fieldNameEta() const { return access()->implementation->fieldNameEta(); }
 
-  /// access the field name used for Phi
-  inline const std::string& fieldNamePhi() const { return access()->implementation->fieldNamePhi(); }
-
   /** \brief Returns a std::vector<double> of the cellDimensions of the given cell ID
-      in natural order of dimensions (dPhi, dEta)
+      in natural order of dimensions (dEta)
 
       Returns a std::vector of the cellDimensions of the given cell ID
       \param cellID is ignored as all cells have the same dimension
-      \return std::vector<double> size 2:
-      -# size in phi
+      \return std::vector<double> size 1:
       -# size in eta
   */
   inline std::vector<double> cellDimensions(const CellID& /*id*/) const {
-    return {access()->implementation->gridSizePhi(), access()->implementation->gridSizeEta()};
+    return {access()->implementation->gridSizeEta()};
   }
 };
 
 } /* End namespace dd4hep                */
-#endif  // DD4HEP_DDCORE_GRIDPHIETA_H
+#endif  // DD4HEP_DDCORE_GRIDETA_H
