+++ detectorSegmentations/include/detectorSegmentations/GridRPhiEtaHandle.h
@@ -1,8 +1,8 @@
-#ifndef DD4HEP_DDCORE_GRIDPHITHETA_H
-#define DD4HEP_DDCORE_GRIDPHITHETA_H 1
+#ifndef DD4HEP_DDCORE_GRIDRPHIETA_H
+#define DD4HEP_DDCORE_GRIDRPHIETA_H 1
 
 // FCCSW
-#include "DetSegmentation/FCCSWGridPhiTheta.h"
+#include "detectorSegmentations/GridRPhiEta.h"
 
 // DD4hep
 #include "DD4hep/Segmentations.h"
@@ -20,9 +20,9 @@ template <typename T>
 class SegmentationWrapper;
 
 /// We need some abbreviation to make the code more readable.
-typedef Handle<SegmentationWrapper<DDSegmentation::FCCSWGridPhiTheta>> FCCSWGridPhiThetaHandle;
+typedef Handle<SegmentationWrapper<DDSegmentation::GridRPhiEta>> GridRPhiEtaHandle;
 
-/// Implementation class for the grid phi-theta segmentation.
+/// Implementation class for the grid phi-eta segmentation.
 /**
  *  Concrete user handle to serve specific needs of client code
  *  which requires access to the base functionality not served
@@ -39,28 +39,30 @@ typedef Handle<SegmentationWrapper<DDSegmentation::FCCSWGridPhiTheta>> FCCSWGrid
  *  The handle base corrsponding to this object in for
  *  conveniance reasons instantiated in DD4hep/src/Segmentations.cpp.
  *
+ *  \author  A. Zaborowska
+ *  \version 1.0
  */
-class FCCSWGridPhiTheta : public FCCSWGridPhiThetaHandle {
+class GridRPhiEta : public GridRPhiEtaHandle {
 public:
   /// Defintiion of the basic handled object
-  typedef FCCSWGridPhiThetaHandle::Object Object;
+  typedef GridRPhiEtaHandle::Object Object;
 
 public:
   /// Default constructor
-  FCCSWGridPhiTheta() = default;
+  GridRPhiEta() = default;
   /// Copy constructor
-  FCCSWGridPhiTheta(const FCCSWGridPhiTheta& e) = default;
+  GridRPhiEta(const GridRPhiEta& e) = default;
   /// Copy Constructor from segmentation base object
-  FCCSWGridPhiTheta(const Segmentation& e) : Handle<Object>(e) {}
+  GridRPhiEta(const Segmentation& e) : Handle<Object>(e) {}
   /// Copy constructor from handle
-  FCCSWGridPhiTheta(const Handle<Object>& e) : Handle<Object>(e) {}
+  GridRPhiEta(const Handle<Object>& e) : Handle<Object>(e) {}
   /// Copy constructor from other polymorph/equivalent handle
   template <typename Q>
-  FCCSWGridPhiTheta(const Handle<Q>& e) : Handle<Object>(e) {}
+  GridRPhiEta(const Handle<Q>& e) : Handle<Object>(e) {}
   /// Assignment operator
-  FCCSWGridPhiTheta& operator=(const FCCSWGridPhiTheta& seg) = default;
+  GridRPhiEta& operator=(const GridRPhiEta& seg) = default;
   /// Equality operator
-  bool operator==(const FCCSWGridPhiTheta& seg) const { return m_element == seg.m_element; }
+  bool operator==(const GridRPhiEta& seg) const { return m_element == seg.m_element; }
   /// determine the position based on the cell ID
   inline Position position(const CellID& id) const { return Position(access()->implementation->position(id)); }
 
@@ -69,49 +71,65 @@ public:
     return access()->implementation->cellID(local, global, volID);
   }
 
-  /// access the grid size in theta
-  inline double gridSizeTheta() const { return access()->implementation->gridSizeTheta(); }
+  /// access the grid size in eta
+  inline double gridSizeEta() const { return access()->implementation->gridSizeEta(); }
 
   /// access the grid size in Phi
   inline int phiBins() const { return access()->implementation->phiBins(); }
 
-  /// access the coordinate offset in theta
-  inline double offsetTheta() const { return access()->implementation->offsetTheta(); }
+  /// access the grid size in R
+  inline double gridSizeR() const { return access()->implementation->gridSizeR(); }
+
+  /// access the coordinate offset in eta
+  inline double offsetEta() const { return access()->implementation->offsetEta(); }
 
   /// access the coordinate offset in Phi
   inline double offsetPhi() const { return access()->implementation->offsetPhi(); }
 
-  /// set the coordinate offset in theta
-  inline void setOffsetTheta(double offset) const { access()->implementation->setOffsetTheta(offset); }
+  /// access the coordinate offset in R
+  inline double offsetR() const { return access()->implementation->offsetR(); }
+
+  /// set the coordinate offset in eta
+  inline void setOffsetEta(double offset) const { access()->implementation->setOffsetEta(offset); }
 
   /// set the coordinate offset in Phi
   inline void setOffsetPhi(double offset) const { access()->implementation->setOffsetPhi(offset); }
 
-  /// set the grid size in theta
-  inline void setGridSizeTheta(double cellSize) const { access()->implementation->setGridSizeTheta(cellSize); }
+  /// set the coordinate offset in R
+  inline void setOffsetR(double offset) const { access()->implementation->setOffsetR(offset); }
+
+  /// set the grid size in eta
+  inline void setGridSizeEta(double cellSize) const { access()->implementation->setGridSizeEta(cellSize); }
 
   /// set the grid size in Phi
   inline void setPhiBins(int cellSize) const { access()->implementation->setPhiBins(cellSize); }
 
-  /// access the field name used for theta
-  inline const std::string& fieldNameTheta() const { return access()->implementation->fieldNameTheta(); }
+  /// set the grid size in R
+  inline void setGridSizeR(double cellSize) const { access()->implementation->setGridSizeR(cellSize); }
+
+  /// access the field name used for eta
+  inline const std::string& fieldNameEta() const { return access()->implementation->fieldNameEta(); }
 
   /// access the field name used for Phi
   inline const std::string& fieldNamePhi() const { return access()->implementation->fieldNamePhi(); }
 
-  /** \brief Returns a std::vector<double> of the cellDimensions of the given cell ID
-      in natural order of dimensions (dPhi, dTheta)
+  /// access the field name used for R
+  inline const std::string& fieldNameR() const { return access()->implementation->fieldNameR(); }
 
+  /** \brief Returns a std::vector<double> of the cellDimensions of the given cell ID
+      in natural order of dimensions (dR, dPhi, dEta)
       Returns a std::vector of the cellDimensions of the given cell ID
       \param cellID is ignored as all cells have the same dimension
-      \return std::vector<double> size 2:
+      \return std::vector<double> size 3:
+      -# size in r
       -# size in phi
-      -# size in theta
+      -# size in eta
   */
   inline std::vector<double> cellDimensions(const CellID& /*id*/) const {
-    return {access()->implementation->gridSizePhi(), access()->implementation->gridSizeTheta()};
+    return {access()->implementation->gridSizeR(), access()->implementation->gridSizePhi(),
+            access()->implementation->gridSizeEta()};
   }
 };
 
 } /* End namespace dd4hep                */
-#endif  // DD4HEP_DDCORE_GRIDPHITHETA_H
+#endif  // DD4HEP_DDCORE_GRIDRPHIETA_H
