+++ detectorSegmentations/src/FCCSWGridPhiEtaHandle.cpp
@@ -1,4 +1,4 @@
-#include "DetSegmentation/FCCSWGridPhiEtaHandle.h"
+#include "detectorSegmentations/FCCSWGridPhiEtaHandle.h"
 #include "DD4hep/detail/Handle.inl"
 
 DD4HEP_INSTANTIATE_HANDLE_UNNAMED(dd4hep::DDSegmentation::FCCSWGridPhiEta);
