+++ detectorSegmentations/src/FCCSWGridPhiThetaHandle.cpp
@@ -1,4 +1,4 @@
-#include "DetSegmentation/FCCSWGridPhiThetaHandle.h"
+#include "detectorSegmentations/FCCSWGridPhiThetaHandle.h"
 #include "DD4hep/detail/Handle.inl"
 
 DD4HEP_INSTANTIATE_HANDLE_UNNAMED(dd4hep::DDSegmentation::FCCSWGridPhiTheta);
