+++ detectorSegmentations/src/GridEtaHandle.cpp
@@ -1,4 +1,4 @@
-#include "DetSegmentation/GridEtaHandle.h"
+#include "detectorSegmentations/GridEtaHandle.h"
 #include "DD4hep/detail/Handle.inl"
 
 DD4HEP_INSTANTIATE_HANDLE_UNNAMED(dd4hep::DDSegmentation::GridEta);
