+++ detectorSegmentations/src/GridRPhiEtaHandle.cpp
@@ -1,4 +1,4 @@
-#include "DetSegmentation/GridRPhiEtaHandle.h"
+#include "detectorSegmentations/GridRPhiEtaHandle.h"
 #include "DD4hep/detail/Handle.inl"
 
 DD4HEP_INSTANTIATE_HANDLE_UNNAMED(dd4hep::DDSegmentation::GridRPhiEta);
