+++ detectorSegmentations/src/GridSimplifiedDriftChamberHandle.cpp
@@ -1,4 +1,4 @@
-#include "DetSegmentation/GridSimplifiedDriftChamberHandle.h"
+#include "detectorSegmentations/GridSimplifiedDriftChamberHandle.h"
 #include "DD4hep/detail/Handle.inl"
 
 DD4HEP_INSTANTIATE_HANDLE_UNNAMED(dd4hep::DDSegmentation::GridSimplifiedDriftChamber);
