+++ detectorSegmentations/src/GridThetaHandle.cpp
@@ -1,4 +1,4 @@
-#include "DetSegmentation/GridThetaHandle.h"
+#include "detectorSegmentations/GridThetaHandle.h"
 #include "DD4hep/detail/Handle.inl"
 
 DD4HEP_INSTANTIATE_HANDLE_UNNAMED(dd4hep::DDSegmentation::GridTheta);
